feat(cli): find and fix AWS instances w/o agents
This commit is the first milestone for Project Capture The Flag. Project
CTF takes a user's cloud credentials, lists all of their hosts,
and installs the agent on hosts without agents. The purpose of
Project CTF is to allow customers to easily close gaps in their agent
deployments by using the CLI to find and install the agent on instances
that the customer would have trouble reaching otherwise.

This commit adds a new subcommand to the Lacework CLI called `ctf`. The
user should invoke this command via `lacework agent ctf`. `ctf` is
currently a no-op and is used as a directory for further subcommands.

This commit adds a subcommand to `ctf` called `aws`. The user
should invoke this command via `lacework agent ctf aws`. `aws` is
currently a no-op and is used as a directory for further subcommands.

This commit adds two subcommands to `aws` called `ec2ic` and `ssh`. The
user should invoke them as either `lacework agent ctf aws ec2ic` or
`lacework agent ctf aws ssh`.

The `ec2ic` subcommand uses EC2InstanceConnect to send SSH keys to the
user's EC2 instances. EC2InstanceConnect is only supported on official
channel AMIs that are Ubuntu 16.04 or later or Amazon Linux 2, so the
customer will not be able to use `ec2ic` for instances running other
AMIs. `ec2ic` adds flags to filter instances on tag, tag key, and region.

The `ssh` subcommand uses existing SSH authentication to install agents
on the user's EC2 instances. `ssh` assumes that all instances that it
finds will have SSH access via the authentication method passed via CLI
flags. `ssh` also adds flags to filter instances on tag, tag key, and
region.

`ec2ic` and `ssh` both have the limitation that they only work for
instances with public IP addresses that are open to the internet on port
22.

Fixes RAIN-37375

Signed-off-by: nschmeller <nick.schmeller@lacework.net>